+++ cper-utils.h
@@ -17,7 +17,7 @@ json_object* bitfield_to_ir(UINT64 bitfield, int num_fields, const char* names[]
 UINT64 ir_to_bitfield(json_object* ir, int num_fields, const char* names[]);
 json_object* uint64_array_to_ir_array(UINT64* array, int len);
 json_object* revision_to_ir(UINT16 revision);
-const char* severity_to_string(UINT8 severity);
+const char* severity_to_string(UINT32 severity);
 void timestamp_to_string(char* out, EFI_ERROR_TIME_STAMP* timestamp);
 void string_to_timestamp(EFI_ERROR_TIME_STAMP* out, const char* timestamp);
 void guid_to_string(char* out, EFI_GUID* guid);
